+++ Source/StateFunding/StateFundingGlobal.cs
@@ -7,6 +7,7 @@ namespace StateFunding
         public static StateFunding fetch;
         public static bool isLoaded;
         public static bool needsDataInit;
+        public static double refRadius = 60000.0;
     }
 }
